// web/src/openapi.gen.ts
export interface components {
    schemas: {
// ...
        NestedAdequacy: {
            /** Trueorder */
            trueOrder: number;
            /** Reducedrejected */
            reducedRejected: boolean;
            /** Overnotpreferredaic */
            overNotPreferredAic: boolean;
            /** Overnotpreferredbic */
            overNotPreferredBic: boolean;
            /** Selectedorderaic */
            selectedOrderAic: number;
            /** Selectedorderbic */
            selectedOrderBic: number;
            /** Recoveredtrueorderaic */
            recoveredTrueOrderAic: boolean;
            /** Recoveredtrueorderbic */
            recoveredTrueOrderBic: boolean;
            reducedVsTrue: components["schemas"]["SelectionStats"];
            trueVsOver: components["schemas"]["SelectionStats"];
        };
// ...
        SelectionStats: {
            /** Lrtstat */
            lrtStat: number;
            /** Lrtp */
            lrtP: number;
            /** Fstat */
            fStat: number;
            /** Fp */
            fP: number;
            /** Daic */
            dAIC: number;
            /** Dbic */
            dBIC: number;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}